requirements.txt
streamlit==1.29.0
pandas==2.1.4
plotly==5.18.0
openpyxl==3.1.2
xlsxwriter==3.2.9
anthropic==0.40.0
python-dotenv==1.0.0